Add files via upload
Here is my best effort at scaling the computation on the mann-kendall trend statistic. It seems to be a solid feature when used in ML models. The computation is: 

For each index, get the total number of points from index:end of series where y is either greater than or less than y.shift(-1), then get a cumulative sum of the differences between both cases. Here is a link for better understanding. 

https://github.com/mmhs013/pyMannKendall/blob/master/pymannkendall/pymannkendall.py

# vectorization approach to calculate mk score, S
def __mk_score(x, n):
    s = 0

    demo = np.ones(n) 
    for k in range(n-1):
        s = s + np.sum(demo[k+1:n][x[k+1:n] > x[k]]) - np.sum(demo[k+1:n][x[k+1:n] < x[k]])
        
    return s